# service/api/template/v1/template_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             (unknown)
// source: api/template/v1/template.proto

package template

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationTemplatev1ServiceDeleteTemplateById = "/api.template.v1.Templatev1Service/DeleteTemplateById"
const OperationTemplatev1ServiceListTemplates = "/api.template.v1.Templatev1Service/ListTemplates"
const OperationTemplatev1ServiceUpsertTemplate = "/api.template.v1.Templatev1Service/UpsertTemplate"

type Templatev1ServiceHTTPServer interface {
	// DeleteTemplateById* Elimina una plantilla por ID.
	//
	// **Descripción**
	// Borra la plantilla indicada por `id`.
	//
	// **Parámetros (query)**
	// - `id` (**obligatorio**): ID interno
	//
	// **Respuestas**
	// - `200 OK`: `meta.code = OK`
	// - `200` con error lógico: `meta.code != OK` (p. ej., ID inexistente)
	//
	// **Ejemplos**
	// `DELETE /templates?id=123`
	DeleteTemplateById(context.Context, *DeleteTemplateByIdRequest) (*DeleteTemplateByIdResponse, error)
	// ListTemplates* Lista todas las plantillas.
	//
	// **Descripción**
	// Devuelve el catálogo completo de plantillas definidas.
	//
	// **Parámetros**
	// _(sin parámetros)_
	//
	// **Respuestas**
	// - `200 OK`: `templates` + `meta.code = OK`
	// - `200` con error lógico: `meta.code != OK` y `meta.message`
	//
	// **Ejemplos**
	// `GET /templates`
	ListTemplates(context.Context, *ListTemplatesRequest) (*ListTemplatesResponse, error)
	// UpsertTemplate* Crea o actualiza una plantilla (upsert).
	//
	// **Descripción**
	// Si `id` es `0` o no se envía → crea; si `id` > 0 → actualiza.
	//
	// **Parámetros (body JSON)**
	// - `id` (opcional, 0=create)
	// - `name` (**obligatorio**)
	//
	// **Respuestas**
	// - `200 OK`: `template` resultante + `meta.code = OK`
	// - `200` con error lógico: validación, duplicidad de `name`, etc.
	//
	// **Ejemplos**
	// `POST /templates`
	// Body: `{ "name":"Template 1" }`
	UpsertTemplate(context.Context, *UpsertTemplateRequest) (*UpsertTemplateResponse, error)
}

func RegisterTemplatev1ServiceHTTPServer(s *http.Server, srv Templatev1ServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/v1/templates", _Templatev1Service_ListTemplates0_HTTP_Handler(srv))
	r.POST("/v1/templates", _Templatev1Service_UpsertTemplate0_HTTP_Handler(srv))
	r.DELETE("/v1/templates", _Templatev1Service_DeleteTemplateById0_HTTP_Handler(srv))
}

func _Templatev1Service_ListTemplates0_HTTP_Handler(srv Templatev1ServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListTemplatesRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTemplatev1ServiceListTemplates)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListTemplates(ctx, req.(*ListTemplatesRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListTemplatesResponse)
		return ctx.Result(200, reply)
	}
}

func _Templatev1Service_UpsertTemplate0_HTTP_Handler(srv Templatev1ServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpsertTemplateRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTemplatev1ServiceUpsertTemplate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpsertTemplate(ctx, req.(*UpsertTemplateRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpsertTemplateResponse)
		return ctx.Result(200, reply)
	}
}

func _Templatev1Service_DeleteTemplateById0_HTTP_Handler(srv Templatev1ServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteTemplateByIdRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTemplatev1ServiceDeleteTemplateById)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteTemplateById(ctx, req.(*DeleteTemplateByIdRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteTemplateByIdResponse)
		return ctx.Result(200, reply)
	}
}

type Templatev1ServiceHTTPClient interface {
	// DeleteTemplateById* Elimina una plantilla por ID.
	//
	// **Descripción**
	// Borra la plantilla indicada por `id`.
	//
	// **Parámetros (query)**
	// - `id` (**obligatorio**): ID interno
	//
	// **Respuestas**
	// - `200 OK`: `meta.code = OK`
	// - `200` con error lógico: `meta.code != OK` (p. ej., ID inexistente)
	//
	// **Ejemplos**
	// `DELETE /templates?id=123`
	DeleteTemplateById(ctx context.Context, req *DeleteTemplateByIdRequest, opts ...http.CallOption) (rsp *DeleteTemplateByIdResponse, err error)
	// ListTemplates* Lista todas las plantillas.
	//
	// **Descripción**
	// Devuelve el catálogo completo de plantillas definidas.
	//
	// **Parámetros**
	// _(sin parámetros)_
	//
	// **Respuestas**
	// - `200 OK`: `templates` + `meta.code = OK`
	// - `200` con error lógico: `meta.code != OK` y `meta.message`
	//
	// **Ejemplos**
	// `GET /templates`
	ListTemplates(ctx context.Context, req *ListTemplatesRequest, opts ...http.CallOption) (rsp *ListTemplatesResponse, err error)
	// UpsertTemplate* Crea o actualiza una plantilla (upsert).
	//
	// **Descripción**
	// Si `id` es `0` o no se envía → crea; si `id` > 0 → actualiza.
	//
	// **Parámetros (body JSON)**
	// - `id` (opcional, 0=create)
	// - `name` (**obligatorio**)
	//
	// **Respuestas**
	// - `200 OK`: `template` resultante + `meta.code = OK`
	// - `200` con error lógico: validación, duplicidad de `name`, etc.
	//
	// **Ejemplos**
	// `POST /templates`
	// Body: `{ "name":"Template 1" }`
	UpsertTemplate(ctx context.Context, req *UpsertTemplateRequest, opts ...http.CallOption) (rsp *UpsertTemplateResponse, err error)
}

type Templatev1ServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewTemplatev1ServiceHTTPClient(client *http.Client) Templatev1ServiceHTTPClient {
	return &Templatev1ServiceHTTPClientImpl{client}
}

// DeleteTemplateById* Elimina una plantilla por ID.
//
// **Descripción**
// Borra la plantilla indicada por `id`.
//
// **Parámetros (query)**
// - `id` (**obligatorio**): ID interno
//
// **Respuestas**
// - `200 OK`: `meta.code = OK`
// - `200` con error lógico: `meta.code != OK` (p. ej., ID inexistente)
//
// **Ejemplos**
// `DELETE /templates?id=123`
func (c *Templatev1ServiceHTTPClientImpl) DeleteTemplateById(ctx context.Context, in *DeleteTemplateByIdRequest, opts ...http.CallOption) (*DeleteTemplateByIdResponse, error) {
	var out DeleteTemplateByIdResponse
	pattern := "/v1/templates"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationTemplatev1ServiceDeleteTemplateById))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListTemplates* Lista todas las plantillas.
//
// **Descripción**
// Devuelve el catálogo completo de plantillas definidas.
//
// **Parámetros**
// _(sin parámetros)_
//
// **Respuestas**
// - `200 OK`: `templates` + `meta.code = OK`
// - `200` con error lógico: `meta.code != OK` y `meta.message`
//
// **Ejemplos**
// `GET /templates`
func (c *Templatev1ServiceHTTPClientImpl) ListTemplates(ctx context.Context, in *ListTemplatesRequest, opts ...http.CallOption) (*ListTemplatesResponse, error) {
	var out ListTemplatesResponse
	pattern := "/v1/templates"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationTemplatev1ServiceListTemplates))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpsertTemplate* Crea o actualiza una plantilla (upsert).
//
// **Descripción**
// Si `id` es `0` o no se envía → crea; si `id` > 0 → actualiza.
//
// **Parámetros (body JSON)**
// - `id` (opcional, 0=create)
// - `name` (**obligatorio**)
//
// **Respuestas**
// - `200 OK`: `template` resultante + `meta.code = OK`
// - `200` con error lógico: validación, duplicidad de `name`, etc.
//
// **Ejemplos**
// `POST /templates`
// Body: `{ "name":"Template 1" }`
func (c *Templatev1ServiceHTTPClientImpl) UpsertTemplate(ctx context.Context, in *UpsertTemplateRequest, opts ...http.CallOption) (*UpsertTemplateResponse, error) {
	var out UpsertTemplateResponse
	pattern := "/v1/templates"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationTemplatev1ServiceUpsertTemplate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}